+++ frontend/tsconfig.json
@@ -26,7 +26,7 @@
     "src/models/s3_bucket.ts",
     "src/models/serverspec.ts",
     "src/models/snapshot.ts",
-    "src/serverspec-gen/arg_list.ts",
+    "src/serverspec-gen/arg_li.ts",
     "src/serverspec-gen/ast-interface.ts",
     "src/serverspec-gen/ast.ts",
     "src/serverspec-gen/it-panel.ts",
